+++ xbalancer-web/src/main/java/org/oopdev/xbalancer/web/global/ProxyConfig.java
@@ -2,8 +2,10 @@ package org.oopdev.xbalancer.web.global;
 
 import io.undertow.server.handlers.proxy.ProxyConnection;
 import org.oopdev.xbalancer.common.util.Strings;
+import org.oopdev.xbalancer.proxy.ProxyProperties;
 import org.oopdev.xbalancer.proxy.ProxyServer;
 import org.oopdev.xbalancer.proxy.client.ProxyClientListener;
+import org.oopdev.xbalancer.proxy.client.ProxyClientRegister;
 import org.oopdev.xbalancer.proxy.host.Host;
 import org.oopdev.xbalancer.service.proxy.ProxyService;
 import org.slf4j.Logger;
@@ -15,6 +17,10 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
 
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
 /**
  * Created by kamilbukum on 03/04/2017.
  */
@@ -24,6 +30,7 @@ import org.springframework.core.annotation.Order;
 @Order(2)
 public class ProxyConfig {
     private static final Logger LOGGER = LoggerFactory.getLogger(ProxyConfig.class);
+
     private ProxyProperties balancer;
 
     public void setBalancer(ProxyProperties balancer) {
@@ -41,29 +48,61 @@ public class ProxyConfig {
     public ProxyServer proxyServer() {
         ProxyServer proxyServer = new ProxyServer(balancer, (exchange, hostMap) -> {
             String requestPath = exchange.getRequestPath();
-            if (requestPath.startsWith("/")) {
+            if (requestPath.equals("/*")) {
+                return hostMap;
+            } else if (requestPath.startsWith("/")) {
                 requestPath = requestPath.substring(1);
             }
+
             if (!Strings.has(requestPath)) {
                 LOGGER.warn("Proxy URL cannot be empty !");
                 return null;
             } else if (requestPath.equals(balancer.getAllPath())) {
                 return hostMap;
             }
-            return proxyService.filter(exchange.getRequestPath(), hostMap);
+            if (!balancer.isDynamic()) {
+                return manuelFilter(requestPath);
+            }
+            return proxyService.filter(requestPath);
         });
+
         proxyServer.getHandler().setListener(new ProxyClientListener() {
             @Override
             public void completed(String name, Host host, ProxyConnection connection) {
-                System.out.println("Dispatch request to " + host);
+                LOGGER.info("Dispatch request to " + host);
             }
 
             @Override
             public void failed(String name, Host host, Exception e) {
-                System.out.println("Dispatch request failed ! Host: " + host);
+                LOGGER.info("Dispatch request failed ! Host: " + host);
             }
         });
         proxyServer.start();
         return proxyServer;
+    }
+
+    public Map<String, Host> manuelFilter(String path) {
+        List<String> hostNames = balancer.getProxies().get(path);
+
+        if (hostNames == null || hostNames.size() == 0) return null;
+
+        Map<String, Host> hostMap = new LinkedHashMap<>();
+        for (Host host : balancer.getHosts()) {
+            if (hostNames.contains(host)) {
+                hostMap.put(host.getName(), host);
+            }
+        }
+        return hostMap;
+    }
+
+    @Bean
+    public ProxyClientRegister clientRegister(ProxyServer proxyServer) {
+        ProxyClientRegister proxyClientRegister = new ProxyClientRegister(proxyServer);
+        if (!balancer.isDynamic()) {
+            for (Host host : balancer.getHosts()) {
+                proxyClientRegister.register(host);
+            }
+        }
+        return proxyClientRegister;
     }
 }